Move BuidlingBuilder.View to BuildingBuilderView more sonar fixes
